pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "summarizer"
version = "0.1.0"
description = "YouTube video summarizer with Whisper and Ollama"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "yt-dlp",
    "openai-whisper",
    "ollama",
    "python-dotenv",
    "requests",
]

[project.scripts]
summarizer = "summarizer.cli:main"
